requirements.txt
psycopg2-binary
python-dotenv
langchain-openai
langchain-core
tabulate
langgraph
pandas
numpy
matplotlib
seaborn
scipy
flask
flask-cors
supabase